server/requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.0.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.16.0
idna==3.10
jwt==1.3.1
kombu==5.5.3
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
more-itertools==10.7.0
passlib==1.7.4
pillow==11.2.1
premailer==3.10.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20

# My Custom MoySklad Client
r-moysklad-api @ git+https://github.com/rinnikha/moysklad_api.git
redis==6.0.0
reportlab==4.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
